requirements.txt
absl-py==1.4.0
altair==5.0.1
altgraph==0.17.4
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
docutils==0.20.1
flatbuffers==23.5.26
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
keyring==24.3.0
libclang==16.0.6
macholib==1.16.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.15
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkginfo==1.9.6
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.16.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
Pympler==1.0.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
readme-renderer==42.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.13.1
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
